+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removed `round` method from the `ServiceContext`. (#366)
 - `ValidatorState` in the `ServiceContext` replaced by the `ValidatorId`. (#366)
 - `add_transaction` in the `ServiceContext` replaced by the `transaction_sender` which implements the `TransactionSend` trait. (#366)
+- The `Node` constructor now requires `db` and `services` variables instead of `blockchain` instance. (#366)
+- The `Blockchain` constructor now requires services keypair and an `ApiSender` instance. (#366)
+- `mount_*_api` methods in `Blockchain` instance now do not require `ApiContext`. (#366)
 
 ### Fixed
 - Fixed `crate_authors!` macro usage, this macro can't return static string in new clap version. (#370)